extract domain constants
