+++ src/features/pokemon-list/domain/constants.ts
@@ -1,0 +1,7 @@
+export const responsiveBreakpoints = {
+  DESKTOP_MIN_WIDTH: 768,
+  TABLET_MIN_WIDTH: 640,
+  DESKTOP_COLUMNS: 5,
+  TABLET_COLUMNS: 3,
+  MOBILE_COLUMNS: 2,
+} as const;
